use double instead of float in CUDA program

+++ cf.h
@@ -4,7 +4,7 @@
 
 const int DIM = 1000;
 const int DIM_SMALL = 200;
-typedef float number;
+typedef double number;
 
 #endif
